refactor(components/items): rename `Item` -> `ItemsListElement`
