+++ app/components/items/index.ts
@@ -1,8 +1,8 @@
 export * from './image'
 export * from './top-card'
 export * from './list'
-export * from './item'
-export * from './item.skeleton'
+export * from './list/element'
+export * from './list/element.skeleton'
 export * from './image.skeleton'
 export * from './list.skeleton'
 export * from './top-card.skeleton'
