+++ app/components/items/list/element.skeleton.tsx
@@ -1,4 +1,4 @@
-import { ItemImageSkeleton } from './image.skeleton'
+import { ItemImageSkeleton } from '../image.skeleton'
 
 import { cn } from '@app/utils/cn'
 import { Skeleton } from '@app/components/ui/skeleton'
